dynamic object version testing
Dynamic Object is a better solution for this case, however, it required a change of settings in Unity Editor to support API. This was known to be unstable, so this code is referenced out to leave it as a possible solution for later.
Current solution is using reflection, which is not only slow but also breaking polymorphism. This should be fixed later on.